Improve work with generics in API
