+++ logging/src/jvmMain/java/io/spine/logging/WithLogging.kt
@@ -26,21 +26,7 @@
 
 package io.spine.logging
 
-import io.kotest.matchers.shouldNotBe
-import io.kotest.matchers.types.shouldBeSameInstanceAs
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Test
-
-@DisplayName("`WithLogging` interface should")
-internal class WithLoggingSpec {
-
-    @Test
-    fun `provide the same logger associated with a class`() {
-        val lc = LoggingConsumer()
-
-        lc.logger shouldNotBe null
-        lc.logger shouldBeSameInstanceAs lc.logger
-    }
-}
-
-private class LoggingConsumer: WithLogging
+/**
+ * Actual interface with provides [logger] with [JvmLogger.Api].
+ */
+public class WithLogging: WithLoggingBase<JvmLogger.Api>
